requirements.txt
aiofiles==0.5.0
astroid==2.4.2
autopep8==1.5.4
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
cryptography==3.1.1
ecdsa==0.14.1
fastapi==0.61.1
h11==0.11.0
isort==5.6.4
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pylint==2.6.0
python-dotenv==0.14.0
python-jose==3.2.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
starlette==0.13.6
toml==0.10.1
uvicorn==0.12.2
wrapt==1.12.1
